split up ImportProjectController from ProjectController
